+++ configs/babel-config/rollup.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  babelHelpers: 'runtime',
+  extensions: ['.mjs', '.cjs', '.js', '.jsx', '.json', '.ts', '.tsx'],
+  rootMode: 'upward',
+};