+++ configs/eslint-config/index.cjs
@@ -91,7 +91,6 @@ module.exports = {
       },
     },
   ],
-  parser: '@babel/eslint-parser',
   plugins: [
     'import',
     'prefer-arrow',